reapply nullable fields patch
