+++ packages/xo-server-sdn-controller/src/index.js
@@ -255,41 +255,6 @@ function isControllerNeeded(xapi) {
   return !(controller?.protocol === PROTOCOL && controller.address === '' && controller.port === 0)
 }
 
-function validateRuleFields(obj, errors, prefix = '') {
-  if (!obj.direction || typeof obj.direction !== 'string') {
-    errors.push(`${prefix}direction is required and must be a string`)
-  }
-  if (!obj.ipRange || typeof obj.ipRange !== 'string') {
-    errors.push(`${prefix}ipRange is required and must be a string`)
-  }
-  if (!obj.protocol || typeof obj.protocol !== 'string') {
-    errors.push(`${prefix}protocol is required and must be a string`)
-  }
-  if (obj.port != null && !Number.isInteger(obj.port)) {
-    errors.push(`${prefix}port must be an integer`)
-  }
-}
-
-function validateRuleWithAllow(obj, errors, prefix = '') {
-  if (obj.allow === undefined || typeof obj.allow !== 'boolean') {
-    errors.push(`${prefix}allow is required and must be a boolean`)
-  }
-  validateRuleFields(obj, errors, prefix)
-}
-
-// Apply a partial update on a rule: a `null` value removes the field
-function applyRulePatch(oldRule, patch) {
-  const rule = { ...oldRule }
-  for (const [key, value] of Object.entries(patch)) {
-    if (value === null) {
-      delete rule[key]
-    } else if (value !== undefined) {
-      rule[key] = value
-    }
-  }
-  return rule
-}
-
 // =============================================================================
 
 class SDNController extends EventEmitter {
@@ -477,340 +442,13 @@ class SDNController extends EventEmitter {
       },
     })
 
-    const createAction = async (res, callback, { sync = false, statusCode = 200, taskProperties }) => {
-      taskProperties.name = 'REST API: ' + taskProperties.name
-      taskProperties.type = 'xo:rest-api:action'
+    this._cleaners.push(this._xo.registerRestRoutes(createRestRoutes(this), '/plugins/sdn-controller'))
 
-      const task = this._xo.tasks.create(taskProperties)
-      const pResult = task.run(callback)
-
-      if (sync) {
-        const result = await pResult
-        res.status(statusCode)
-        return result ?? {}
-      } else {
-        pResult.catch(() => {})
-        res.status(202).set('Location', `/rest/v0/tasks/${task.id}`).json({ taskId: task.id })
-        return undefined
-      }
-    }
-
-    this._xo.registerRestApi(
-      {
-        networks: {
-          ':id/actions/add_traffic_rule': {
-            _post: async (req, res, next) => {
-              const validationErrors = []
-              validateRuleWithAllow(req.body, validationErrors)
-              if (validationErrors.length > 0) {
-                throw invalidParameters(validationErrors)
-              }
-
-              const rule = {
-                allow: req.body.allow,
-                direction: req.body.direction,
-                ipRange: req.body.ipRange,
-                protocol: req.body.protocol,
-                networkId: req.params.id,
-              }
-
-              if (req.body.port != null) {
-                rule.port = req.body.port
-              }
-
-              return createAction(res, () => this._addNetworkRule(rule), {
-                sync: req.query.sync ?? false,
-                statusCode: 204,
-                taskProperties: {
-                  name: 'add network traffic rule',
-                  objectId: rule.networkId,
-                },
-              })
-            },
-          },
-          ':id/actions/delete_traffic_rule': {
-            _post: async (req, res, next) => {
-              const validationErrors = []
-              validateRuleFields(req.body, validationErrors)
-              if (validationErrors.length > 0) {
-                throw invalidParameters(validationErrors)
-              }
-
-              const rule = {
-                direction: req.body.direction,
-                ipRange: req.body.ipRange,
-                protocol: req.body.protocol,
-                networkId: req.params.id,
-              }
-
-              if (req.body.port != null) {
-                rule.port = req.body.port
-              }
-
-              return createAction(res, () => this._deleteNetworkOfRule(rule), {
-                sync: req.query.sync ?? false,
-                statusCode: 204,
-                taskProperties: {
-                  name: 'delete network traffic rule',
-                  objectId: rule.networkId,
-                },
-              })
-            },
-          },
-        },
-        vifs: {
-          ':id/actions/add_traffic_rule': {
-            _post: async (req, res, next) => {
-              const validationErrors = []
-              validateRuleWithAllow(req.body ?? {}, validationErrors)
-              if (validationErrors.length > 0) {
-                throw invalidParameters(validationErrors)
-              }
-
-              const rule = {
-                allow: req.body.allow,
-                direction: req.body.direction,
-                ipRange: req.body.ipRange,
-                protocol: req.body.protocol,
-                vifId: req.params.id,
-              }
-
-              if (req.body.port != null) {
-                rule.port = req.body.port
-              }
-
-              return createAction(res, () => this._addRule(rule), {
-                sync: req.query.sync ?? false,
-                statusCode: 204,
-                taskProperties: {
-                  name: 'add vif traffic rule',
-                  objectId: rule.vifId,
-                },
-              })
-            },
-          },
-          ':id/actions/delete_traffic_rule': {
-            _post: async (req, res, next) => {
-              const validationErrors = []
-              validateRuleFields(req.body, validationErrors)
-              if (validationErrors.length > 0) {
-                throw invalidParameters(validationErrors)
-              }
-
-              const rule = {
-                direction: req.body.direction,
-                ipRange: req.body.ipRange,
-                protocol: req.body.protocol,
-                vifId: req.params.id,
-              }
-
-              if (req.body.port != null) {
-                rule.port = req.body.port
-              }
-
-              return createAction(res, () => this._deleteRule(rule), {
-                sync: req.query.sync ?? false,
-                statusCode: 204,
-                taskProperties: {
-                  name: 'delete vif traffic rule',
-                  objectId: rule.vifId,
-                },
-              })
-            },
-          },
-        },
-      },
-      '/plugins/sdn-controller'
-    )
-    this._cleaners.push(
-      this._xo.registerRestRoutes(
-        [
-          {
-            endpoint: '/networks/{id}/actions/update_traffic_rule',
-            description:
-              'Update a rule on a network: `oldRule` identifies the rule to update and must be given in full, `newRule` is a partial update where a field set to `null` is removed from the rule.\n\nRequired privilege:\n - resource: network, action: update:other_config',
-
-            method: 'post',
-            tags: ['sdn-controller'],
-            params: {
-              id: { type: 'string', example: 'b97f4e69-d275-4b25-9dc9-c1ac4e9b3fa5' },
-            },
-            query: {
-              sync: { type: 'boolean', optional: true },
-            },
-            body: {
-              oldRule: {
-                type: 'object',
-                fields: {
-                  allow: { type: 'boolean', example: true },
-                  direction: { type: 'string', example: 'to' },
-                  ipRange: { type: 'string', example: '192.168.0.0/24' },
-                  protocol: { type: 'string', example: 'tcp' },
-                  port: { type: 'number', example: 443, optional: true },
-                },
-              },
-              newRule: {
-                type: 'object',
-                fields: {
-                  allow: { type: 'boolean', example: true, optional: true },
-                  direction: { type: 'string', example: 'to', optional: true },
-                  ipRange: { type: 'string', example: '10.0.0.0/8', optional: true },
-                  protocol: { type: 'string', example: 'tcp', optional: true },
-                  port: { type: 'number', example: 80, optional: true, nullable: true },
-                },
-              },
-            },
-            responses: [
-              {
-                status: 204,
-                description: 'Rule updated successfully',
-              },
-              {
-                status: 404,
-                description: 'Old rule does not exist on this network',
-              },
-            ],
-            middlewares: [
-              { name: 'json' },
-              { name: 'acl', acls: { resource: 'network', action: 'update:other_config', objectId: 'params.id' } },
-            ],
-            callback: ({ req, createAction }) => {
-              return createAction(
-                async () => {
-                  const { oldRule, newRule: partialNewRule } = req.body
-                  const networkId = req.params.id
-                  const network = this._xo.getObject(networkId, 'network')
-                  const networkRules = JSON.parse(network.other_config[SDN_CONTROLLER_OF_RULES_KEY] || '[]').map(
-                    JSON.parse
-                  )
-                  if (
-                    !networkRules.some(
-                      rule =>
-                        rule.allow === oldRule.allow &&
-                        rule.direction === oldRule.direction &&
-                        rule.ipRange === oldRule.ipRange &&
-                        rule.port === oldRule.port &&
-                        rule.protocol.toLowerCase() === oldRule.protocol.toLowerCase()
-                    )
-                  ) {
-                    throw noSuchObject(JSON.stringify(oldRule), 'traffic-rule')
-                  }
-                  const newRule = applyRulePatch(oldRule, partialNewRule)
-
-                  await this._deleteNetworkOfRule({ ...oldRule, networkId })
-                  await this._addNetworkRule({ ...newRule, networkId })
-                },
-                {
-                  sync: req.query.sync ?? false,
-                  statusCode: 204,
-                  taskProperties: {
-                    name: 'update network traffic rule',
-                    objectId: req.params.id,
-                    params: req.body,
-                    objectType: 'network',
-                  },
-                }
-              )
-            },
-          },
-          {
-            endpoint: '/vifs/{id}/actions/update_traffic_rule',
-            description:
-              'Update a rule on a VIF: `oldRule` identifies the rule to update and must be given in full, `newRule` is a partial update where a field set to `null` is removed from the rule.\n\nRequired privilege:\n - resource: vif, action: update:other_config',
-            method: 'post',
-            tags: ['sdn-controller'],
-            params: {
-              id: { type: 'string', example: 'b97f4e69-d275-4b25-9dc9-c1ac4e9b3fa5' },
-            },
-            query: {
-              sync: { type: 'boolean', optional: true },
-            },
-            body: {
-              oldRule: {
-                type: 'object',
-                fields: {
-                  allow: { type: 'boolean', example: true },
-                  direction: { type: 'string', example: 'from' },
-                  ipRange: { type: 'string', example: '192.168.0.0/24' },
-                  protocol: { type: 'string', example: 'tcp' },
-                  port: { type: 'number', example: 443, optional: true },
-                },
-              },
-              newRule: {
-                type: 'object',
-                fields: {
-                  allow: { type: 'boolean', example: true, optional: true },
-                  direction: { type: 'string', example: 'to', optional: true },
-                  ipRange: { type: 'string', example: '10.0.0.0/8', optional: true },
-                  protocol: { type: 'string', example: 'tcp', optional: true },
-                  port: { type: 'number', example: 80, optional: true, nullable: true },
-                },
-              },
-            },
-
-            responses: [
-              {
-                status: 204,
-                description: 'Rule updated successfully',
-              },
-              {
-                status: 404,
-                description: 'Old rule does not exist on this VIF',
-              },
-            ],
-            middlewares: [
-              { name: 'json' },
-              { name: 'acl', acls: { resource: 'vif', action: 'update:other_config', objectId: 'params.id' } },
-            ],
-            callback: ({ req, createAction }) => {
-              return createAction(
-                async () => {
-                  const { oldRule, newRule: partialNewRule } = req.body
-                  const vifId = req.params.id
-                  const vif = this._xo.getObject(vifId, 'VIF')
-                  const rawVifRules = vif.other_config[SDN_CONTROLLER_OF_RULES_KEY]
-                  const vifRules = rawVifRules !== undefined ? JSON.parse(rawVifRules).map(JSON.parse) : []
-                  if (
-                    !vifRules.some(
-                      rule =>
-                        rule.allow === oldRule.allow &&
-                        rule.direction === oldRule.direction &&
-                        rule.ipRange === oldRule.ipRange &&
-                        rule.port === oldRule.port &&
-                        rule.protocol.toLowerCase() === oldRule.protocol.toLowerCase()
-                    )
-                  ) {
-                    throw noSuchObject(JSON.stringify(oldRule), 'traffic-rule')
-                  }
-
-                  const newRule = applyRulePatch(oldRule, partialNewRule)
-
-                  await this._deleteRule({ ...oldRule, vifId })
-                  await this._addRule({ ...newRule, vifId })
-                },
-                {
-                  sync: req.query.sync ?? false,
-                  statusCode: 204,
-                  taskProperties: {
-                    name: 'update vif traffic rule',
-                    objectId: req.params.id,
-                    params: req.body,
-                    objectType: 'VIF',
-                  },
-                }
-              )
-            },
-          },
-        ],
-        '/plugins/sdn-controller'
-      )
-    )
     forOwn(this._xo.getAllXapis(), xapi => {
       if (xapi.status === 'connected') {
         this._handleConnectedXapi(xapi)
       }
     })
-    this._cleaners.push(this._xo.registerRestRoutes(createRestRoutes(this), '/plugins/sdn-controller'))
 
     const handleConnectedServer = ({ xapi }) => this._handleConnectedXapi(xapi)
     const handleDisconnectedServer = ({ xapi }) => this._handleDisconnectedXapi(xapi)
